pyproject.toml
[tool.poetry]
name = "antifrod_analyze"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
